+++ src/app/app-routing.module.ts
@@ -14,7 +14,10 @@ import {
   InstitutionListComponent,
   AddUserComponent,
   AddInstitutionComponent,
+  AddMartyrComponent,
+  MartyrListComponent,
   UserListComponent,
+  ContactComponent,
   MartyrInfoComponent
 } from './pages';
 
@@ -22,7 +25,22 @@ const routes: Routes = [
   {
     path: '',
     component: ClientLayoutComponent,
-    children: [{ path: '', component: HomepageComponent }],
+    children: [
+      { path: '', component: HomepageComponent },
+      { path: 'contact', component: ContactComponent },
+      {
+        path: '',
+        component: HomepageComponent
+      },
+      {
+        path: 'martyr/:MartyrID',
+        component: MartyrInfoComponent,
+        data: {
+          title: 'Martyr Information',
+          icon: 'fa fa-2x fa-user-plus',
+        },
+      },
+    ],
   },
   {
     path: 'admin',
@@ -53,7 +71,7 @@ const routes: Routes = [
         path: 'user/edit/:UserID',
         component: AddUserComponent,
         data: {
-          title: 'User Edit',
+          title: 'Edit User Information',
           icon: 'fa fa-2x fa-user-edit',
           authorize: [Roles.Root, Roles.Administrator, Roles.InstitutionAdmin],
         },
@@ -80,7 +98,7 @@ const routes: Routes = [
         path: 'institution/edit/:InstitutionID',
         component: AddInstitutionComponent,
         data: {
-          title: 'Institution Edit',
+          title: 'Edit Institution Information',
           icon: 'fa fa-2x fa-edit',
           authorize: [Roles.Root, Roles.Administrator],
         },
@@ -94,6 +112,21 @@ const routes: Routes = [
           authorize: [Roles.Root, Roles.Administrator, Roles.InstitutionAdmin],
         },
       },
+      {
+        path: 'martyrs',
+        component: MartyrListComponent,
+        data: { title: 'Martyr List', icon: 'fa-2x fas fa-landmark' },
+      },
+      {
+        path: 'martyr/add',
+        component: AddMartyrComponent,
+        data: { title: 'Martyr Add', icon: 'fas fa-monument fa-2x' },
+      },
+      {
+        path: 'martyr/edit/:MartyrID',
+        component: AddMartyrComponent,
+        data: { title: 'Edit Martyr Information', icon: 'fas fa-monument fa-2x' },
+      },
     ],
   },
   { path: 'login', component: LoginComponent },
@@ -103,5 +136,5 @@ const routes: Routes = [
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
 export const routingComponents = [];