+++ src/app/components/layouts/admin/admin-header/admin-header.component.ts
@@ -1,15 +1,40 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ChangePasswordModalComponent } from 'src/app/components/change-password-modal/change-password-modal.component';
+import { LanguageService, AuthService } from '../../../../utils';
 
 @Component({
   selector: 'app-admin-header',
   templateUrl: './admin-header.component.html',
-  styleUrls: ['./admin-header.component.scss']
+  styleUrls: ['./admin-header.component.scss'],
 })
 export class AdminHeaderComponent implements OnInit {
+  constructor(
+    private _languageService: LanguageService,
+    private _authService: AuthService,
+    private _dialog:MatDialog
+  ) {}
 
-  constructor() { }
-
-  ngOnInit(): void {
+  examineOpenDialog() {
+    this._dialog.open(ChangePasswordModalComponent, {
+      width: '400px',
+    });
   }
 
+  userInformation = this._authService.currentUserValue.result;
+  lang: string =
+    this._languageService.getLanguage() == 'en'
+      ? 'us'
+      : this._languageService.getLanguage() || 'tr';
+  ngOnInit(): void {}
+
+  setLang(lang: string) {
+    this.lang = lang == 'en' ? 'us' : lang;
+    this._languageService.setLanguage(lang);
+    window.location.reload();
+  }
+
+  async signout() {
+    await this._authService.logout();
+  }
 }