+++ src/app/components/layouts/admin/admin-layout/admin-layout.module.ts
@@ -13,18 +13,33 @@ import { FormsModule } from '@angular/forms';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatIconModule } from '@angular/material/icon';
 import { MatDialogModule } from '@angular/material/dialog';
-import {InstitutionInfoDialogComponent,UserInfoDialogComponent,MartyrInfoDialogComponent} from '../../';
+import { MatMomentDateModule } from '@angular/material-moment-adapter';
+import { MatMenuModule } from '@angular/material/menu';
+import { NgSearchFilterModule } from 'ng-search-filter';
+import { CKEditorModule } from 'ckeditor4-angular';
+import {
+  InstitutionInfoDialogComponent,
+  UserInfoDialogComponent,
+  DialogWindowComponent,
+  ChangePasswordModalComponent,
+  MartyrImageDialogComponent,
+  MartyrInfoDialogComponent
+} from '../../';
 import {
   DashboardComponent,
   LoginComponent,
   AddUserComponent,
   AddInstitutionComponent,
   InstitutionListComponent,
+  AddMartyrComponent,
+  MartyrListComponent,
   UserListComponent,
   MartyrInfoComponent
 } from '../../../../pages/admin';
+import { MatNativeDateModule } from '@angular/material/core';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http);
@@ -33,29 +48,40 @@ export function HttpLoaderFactory(http: HttpClient) {
 @NgModule({
   declarations: [
     AdminLayoutComponent,
-    DashboardComponent,
     AdminFooterComponent,
     AdminHeaderComponent,
+    MartyrImageDialogComponent,
     AdminControlSidebarComponent,
     AdminSidebarComponent,
+    DashboardComponent,
     LoginComponent,
     AddUserComponent,
     AddInstitutionComponent,
     InstitutionListComponent,
+    AddMartyrComponent,
+    MartyrListComponent,
     UserListComponent,
-    UserInfoDialogComponent,
+    DialogWindowComponent,
+    ChangePasswordModalComponent,
     InstitutionInfoDialogComponent,
     MartyrInfoDialogComponent,
-    MartyrInfoComponent
+    MartyrInfoComponent,
+    UserInfoDialogComponent,
   ],
-  entryComponents: [UserInfoDialogComponent, InstitutionInfoDialogComponent],
   imports: [
     CommonModule,
     RouterModule,
     FormsModule,
     MatSnackBarModule,
+    MatDialogModule,
     MatInputModule,
     MatFormFieldModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    CKEditorModule,
+    MatMomentDateModule,
+    MatMenuModule,
+    NgSearchFilterModule,
     MatIconModule,
     MatDialogModule,
     TranslateModule.forRoot({
@@ -68,4 +94,4 @@ export function HttpLoaderFactory(http: HttpClient) {
   ],
   providers: [],
 })
-export class AdminLayoutModule {}
+export class AdminLayoutModule { }