+++ src/app/pages/admin/add-martyr/add-martyr.model.ts
@@ -1,0 +1,16 @@
+export class Martyr {
+  MartyrFirstName: string;
+  MartyrLastName: string;
+  MartyrDateOfBirth: Date;
+  MartyrDateOfDeath: Date;
+  MartyrCity: string;
+  MartyrDistrict: string;
+  MartyrPlaceOfDeath: string;
+  MartyrContent: string;
+  Image: File;
+  RankName: string;
+}
+
+export class Rank {
+  RankName: string;
+}