+++ src/app/pages/admin/index.ts
@@ -5,3 +5,5 @@ export * from './user-list/user-list.component'
 export * from './add-institution/add-institution.component';
 export * from './institution-list/institution-list.component';
 export * from './martyr-info/martyr-info.component';
+export * from './add-martyr/add-martyr.component';
+export * from './martyr-list/martyr-list.component';