+++ src/app/pages/admin/user-list/user-list.component.ts
@@ -4,23 +4,26 @@ import { UserService } from '../../../utils/services';
 import { TranslateService } from '@ngx-translate/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatDialog } from '@angular/material/dialog';
-import { UserInfoDialogComponent } from '../../../components';
+import {
+  DialogWindowComponent,
+  UserInfoDialogComponent,
+} from '../../../components/';
 
 @Component({
   selector: 'app-user-list',
   templateUrl: './user-list.component.html',
-  styleUrls: ['./user-list.component.scss']
+  styleUrls: ['./user-list.component.scss'],
 })
 export class UserListComponent implements OnInit {
-
   constructor(
     private _userService: UserService,
     private _snackBar: MatSnackBar,
     private _translateService: TranslateService,
     private _dialog: MatDialog
-  ) { }
+  ) {}
 
   data: User[];
+  searchText: string;
 
   async ngOnInit() {
     this.data = <User[]>await this._userService.listAsync();
@@ -31,50 +34,59 @@ export class UserListComponent implements OnInit {
       message: '',
       panelClass: 'notification__success',
     };
-    try {
-      await this._userService.deleteAsync({ UserID });
-      this.data.splice(
-        this.data.findIndex(
-          (user) => user.UserID == UserID
-        ),
-        1
-      );
-      this._translateService
-        .get('Institution information was successfully deleted')
-        .subscribe((value) => (notification.message = value));
-    } catch (error) {
-      console.log(error);
-      notification.panelClass = 'notification__error';
-      switch (error.status) {
-        case 401:
+    const diologRef = this._dialog.open(DialogWindowComponent, {
+      data: {
+        message: 'Are you sure you want to delete the user ?',
+        icon: 'fa fa-exclamation',
+      },
+    });
+
+    diologRef.afterClosed().subscribe(async (result: boolean) => {
+      if (result) {
+        try {
+          await this._userService.deleteAsync({ UserID });
+          this.data.splice(
+            this.data.findIndex((user) => user.UserID == UserID),
+            1
+          );
           this._translateService
-            .get('Unauthorized transaction !')
+            .get('Institution information was successfully deleted')
             .subscribe((value) => (notification.message = value));
-          break;
-        case 417:
-          this._translateService
-            .get('Please enter correct institution information !')
-            .subscribe((value) => (notification.message = value));
-          break;
-        case 407:
-          window.location.reload();
-          break;
-        default:
-          this._translateService
-            .get(
-              'Server error occurred, please try again later If the error persists, we ask you to report this to the authorities'
-            )
-            .subscribe((value) => (notification.message = value));
-          break;
+        } catch (error) {
+          console.log(error);
+          notification.panelClass = 'notification__error';
+          switch (error.status) {
+            case 401:
+              this._translateService
+                .get('Unauthorized transaction !')
+                .subscribe((value) => (notification.message = value));
+              break;
+            case 417:
+              this._translateService
+                .get('Please enter correct institution information !')
+                .subscribe((value) => (notification.message = value));
+              break;
+            case 407:
+              window.location.reload();
+              break;
+            default:
+              this._translateService
+                .get(
+                  'Server error occurred, please try again later If the error persists, we ask you to report this to the authorities'
+                )
+                .subscribe((value) => (notification.message = value));
+              break;
+          }
+        } finally {
+          this._snackBar.open(notification.message, 'X', {
+            duration: 3000,
+            panelClass: notification.panelClass,
+            verticalPosition: 'bottom',
+            horizontalPosition: 'right',
+          });
+        }
       }
-    } finally {
-      this._snackBar.open(notification.message, 'X', {
-        duration: 3000,
-        panelClass: notification.panelClass,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'right',
-      });
-    }
+    });
   }
 
   examineOpenDialog(UserID) {