+++ src/app/pages/client/homepage/homepage.model.ts
@@ -1,0 +1,11 @@
+export class Martyr {
+  MartyrFirstName: string;
+  MartyrLastName: string;
+  MartyrCity: string;
+  MartyrDistrict: string;
+  MartyrDateOfBirth: string;
+  MartyrDateOfDeath: string;
+  MartyrImagePath: string;
+  RankAbbreviation: string;
+  MartyrPlaceOfDeath: string;
+}