+++ src/app/utils/services/auth/auth.service.ts
@@ -156,4 +156,18 @@ export class AuthService {
     }
     return passwordArray.join('');
   }
+
+  async logout() {
+    localStorage.removeItem('currentUser');
+    this._router.navigate(['login']);
+  }
+
+  async changePassword(values) {
+    return await this._apiFetchService.requestAsync(
+      'PUT',
+      'change-password',
+      values,
+      true
+    );
+  }
 }