+++ src/app/utils/services/index.ts
@@ -2,3 +2,5 @@ export * from './api-fetch/api-fetch.service';
 export * from './auth/auth.service';
 export * from './institution/institution.service';
 export * from './user/user.service';
+export * from './martyr/martyr.service';
+export * from './rank/rank.service';